extract code into AbstractReplicator
this is the common code between open-replicator and shyko's library.
