+++ src/main/java/com/zendesk/maxwell/replication/TableCache.java
@@ -11,12 +11,9 @@ import com.zendesk.maxwell.schema.Table;
 public class TableCache {
 	private final HashMap<Long, Table> tableMapCache = new HashMap<>();
 	private final HashMap<Long, String> blacklistedTableCache = new HashMap<>();
-	// open-replicator keeps a very similar cache, but we can't get access to it.
-	public void processEvent(Schema schema, MaxwellFilter filter, TableMapEvent event) {
-		Long tableId = event.getTableId();
+
+	public void processEvent(Schema schema, MaxwellFilter filter, Long tableId, String dbName, String tblName) {
 		if ( !tableMapCache.containsKey(tableId) ) {
-			String dbName = new String(event.getDatabaseName().getValue());
-			String tblName = new String(event.getTableName().getValue());
 			Database db = schema.findDatabase(dbName);
 
 			if ( filter != null && filter.isTableBlacklisted(dbName, tblName) )
@@ -32,6 +29,15 @@ public class TableCache {
 					tableMapCache.put(tableId, tbl);
 			}
 		}
+
+	}
+
+	// open-replicator keeps a very similar cache, but we can't get access to it.
+	public void processEvent(Schema schema, MaxwellFilter filter, TableMapEvent event) {
+		String dbName = new String(event.getDatabaseName().getValue());
+		String tblName = new String(event.getTableName().getValue());
+
+		processEvent(schema, filter, event.getTableId(), dbName, tblName);
 	}
 
 	public Table getTable(Long tableId) {
